+++ core/src/components/cat-menu/cat-menu.tsx
@@ -27,6 +27,11 @@ export class CatMenu {
   @Prop() placement: Placement = 'bottom-start';
 
   /**
+   * Do not close the menu on outside clicks.
+   */
+  @Prop() noAutoClose = false;
+
+  /**
    * Emitted when the menu is opened.
    */
   @Event() catOpen!: EventEmitter<FocusEvent>;
@@ -38,8 +43,13 @@ export class CatMenu {
 
   @Listen('catClick')
   clickHandler(event: CustomEvent<MouseEvent>) {
+    if (!this.trigger) {
+      this.initTrigger();
+      this.show();
+    }
+
     // hide menu on button click
-    if (this.content && event.composedPath().includes(this.content)) {
+    if (!this.noAutoClose && this.content && event.composedPath().includes(this.content)) {
       this.close();
     }
   }
@@ -54,16 +64,7 @@ export class CatMenu {
   }
 
   componentDidLoad(): void {
-    this.trigger = this.findTrigger();
-    this.trigger?.setAttribute('aria-haspopup', 'true');
-    this.trigger?.setAttribute('aria-expanded', 'false');
-    this.trigger?.setAttribute('aria-controls', this.contentId);
-    this.content?.setAttribute('id', this.contentId);
-    if (this.trigger && this.content) {
-      this.trigger?.addEventListener('click', () => this.show());
-      autoUpdate(this.trigger, this.content, () => this.update());
-    }
-
+    this.initTrigger();
     this.keyListener = event => {
       if (this.content && ['ArrowDown', 'ArrowUp'].includes(event.key)) {
         const targetElements = tabbable(this.content, { includeContainer: false, getShadowRoot: true });
@@ -99,6 +100,20 @@ export class CatMenu {
     return `cat-menu-${this.id}`;
   }
 
+  private initTrigger() {
+    this.trigger = this.findTrigger();
+    this.trigger?.setAttribute('aria-haspopup', 'true');
+    this.trigger?.setAttribute('aria-expanded', 'false');
+    this.trigger?.setAttribute('aria-controls', this.contentId);
+    this.content?.setAttribute('id', this.contentId);
+    if (this.trigger && this.content) {
+      this.trigger?.addEventListener('click', () => {
+        this.trap?.active ? this.close() : this.show();
+      });
+      autoUpdate(this.trigger, this.content, () => this.update());
+    }
+  }
+
   private show() {
     if (this.content) {
       this.content.style.display = 'block';
@@ -111,8 +126,11 @@ export class CatMenu {
               getShadowRoot: true
             },
             allowOutsideClick: true,
-            clickOutsideDeactivates: event => !this.content || !event.composedPath().includes(this.content),
-            onPostDeactivate: () => this.hide()
+            clickOutsideDeactivates: event =>
+              !this.noAutoClose &&
+              (!this.content || !event.composedPath().includes(this.content)) &&
+              (!this.trigger || !event.composedPath().includes(this.trigger)),
+            onPostDeactivate: () => this.close()
           });
       this.trap.activate();
     }